refactor(core): start split the core_api file
