+++ core/src/api/run_manager.rs
@@ -1,0 +1,8 @@
+use crate::{app, run};
+
+/// API State
+
+pub(crate) struct RunManager {
+    pub(crate) pending_apps: Vec<(app::App, run::Credentials, run::Secrets, bool)>,
+    pub(crate) pending_runs: Vec<String>,
+}
