+++ core/src/lib.rs
@@ -173,3 +173,9 @@ pub mod open_telemetry;
 pub mod otel_log_format;
 
 pub mod mem_check;
+
+pub mod api {
+    pub mod api_state;
+    pub mod projects;
+    pub(crate) mod run_manager;
+}
